Allow interactive tooltips, use doTimeout
Leave tooltip open when hovered,
use https://github.com/cowboy/jquery-dotimeout instead of setTimeout:
 - better user experience when debouncing redundant events,
 - avoid excessive ajax requests due to redundant events.

NOTE: with the default animation you can stop the tooltip disappearing by hovering it even if the hiding animation started. but i couldn't get it to work with a user-selected animation, and anyway it seems to cause a strange event-calling sequence :-/. So except for the default fade in/out, hiding will only be canceled if you hover quickly, whereas once it started it will not be stopped.